backend/requirements.txt
alembic>=1.13.0
bcrypt>=4.0.0
email-validator>=2.0.0
fastapi>=0.128.0
psycopg2-binary>=2.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-jose>=3.3.0
python-multipart>=0.0.9
resend>=2.0.0
slowapi>=0.1.9
SQLAlchemy>=2.0.0
starlette>=0.36.0
uvicorn>=0.30.0
